Implement validation for category delete

+++ src/Controller/admin/CategoriesController.php
@@ -80,7 +80,11 @@ class CategoriesController extends AbstractController
     public function delete(Request $request, Category $category, CategoryRepository $categoryRepository): Response
     {
         if ($this->isCsrfTokenValid('delete'.$category->getId(), $request->request->get('_token'))) {
-            $categoryRepository->remove($category);
+            if ($category->getPosts()->count() === 0 ) {
+                $categoryRepository->remove($category);
+            } else {
+                $this->addFlash('notice', 'There are posts under this category.');
+            }
         }
 
         return $this->redirectToRoute('app_categories_index', [], Response::HTTP_SEE_OTHER);
